Fixed the broken auto* build tools
Fixed autoconf, automake, libtool and other tools designed
to make the build process easier; I question that though.

+++ honeyd/config.c
@@ -911,7 +911,7 @@ template_print(struct evbuffer *buffer, struct template *tmpl)
 		evbuffer_add_printf(buffer, "  ethernet address: %s\n",
 		    addr_ntoa(tmpl->ethernet_addr));
 	evbuffer_add_printf(buffer, "  IP id: %u\n", tmpl->id);
-	evbuffer_add_printf(buffer, "  TCP seq: %lx\n", tmpl->seq);
+	evbuffer_add_printf(buffer, "  TCP seq: %lx\n", (long unsigned int)tmpl->seq);
 	evbuffer_add_printf(buffer, "  TCP drop: in: %d syn: %d\n",
 	    tmpl->drop_inrate, tmpl->drop_synrate);
 	evbuffer_add_printf(buffer, "  refcnt: %d\n", tmpl->refcnt);
